[CL-2101] Moved SideFxUserService patch to core
